feat: 구인 공고 필터·모달 뷰 분리 및 선택 상태 zustand 스토어로 관리
