+++ src/app/routes/index.tsx
@@ -6,8 +6,8 @@ import Login from '@/pages/login'
 import AdminDashboardPage from '@/pages/members/dashboard'
 import AdminUserManagementPage from '@/pages/members/users'
 import AdminWithdrawalManagementPage from '@/pages/members/withdrawals'
-import RecruitmentApplicationsPage from '@/pages/recruitment/applications'
-import RecruitmentPostManagementPage from '@/pages/recruitment/posts'
+import RecruitmentPostManagementPage from '@/pages/recruitment/applications'
+import RecruitmentApplicationsPage from '@/pages/recruitment/recruitment'
 import LectureManagementPage from '@/pages/study/courses'
 import StudyGroupManagementPage from '@/pages/study/groups'
 import ReviewManagementPage from '@/pages/study/reviews'
