--- src/pages/recruitment/posts/index.tsx
@@ -1,5 +1,0 @@
-// 스터디 구인 / 지원내역 관리
-
-export default function RecruitmentPostManagementPage() {
-  return <>스터디 구인 / 지원내역 관리</>
-}
