+++ tsconfig.json
@@ -10,5 +10,5 @@
       "@mocks/*": ["src/mocks/*"]
     }
   },
-  "include": ["src", "vite.config.ts"]
+  "include": ["src", "vite.config.ts", "src/store/.ts"]
 }
